+++ src/EditorFeatures/Core/TaskList/TaskListUpdatedArgs.cs
@@ -6,7 +6,7 @@ using System.Collections.Immutable;
 using Microsoft.CodeAnalysis.Common;
 using Microsoft.CodeAnalysis.TodoComments;
 
-namespace Microsoft.CodeAnalysis.Editor
+namespace Microsoft.CodeAnalysis.TaskList
 {
     internal sealed class TaskListUpdatedArgs : UpdatedEventArgs
     {
